rename idev-tool-state
