+++ src/Devtools/dev-tool-initializer.ts
@@ -3,7 +3,7 @@
 import { DevToolsChromeAdapter } from 'background/dev-tools-chrome-adapter';
 import { StoreProxy } from '../common/store-proxy';
 import { StoreNames } from '../common/stores/store-names';
-import { DevToolStoreData } from '../common/types/store-data/idev-tool-state';
+import { DevToolStoreData } from '../common/types/store-data/dev-tool-store-data';
 import { InspectHandler } from './inspect-handler';
 
 export class DevToolInitializer {
