+++ src/Devtools/inspect-handler.ts
@@ -4,7 +4,7 @@ import { DevToolsChromeAdapter } from 'background/dev-tools-chrome-adapter';
 import { BaseStore } from '../common/base-store';
 import { ConnectionNames } from '../common/constants/connection-names';
 import { DevToolsOpenMessage } from '../common/types/dev-tools-open-message';
-import { DevToolStoreData } from '../common/types/store-data/idev-tool-state';
+import { DevToolStoreData } from '../common/types/store-data/dev-tool-store-data';
 
 export class InspectHandler {
     private devToolsStore: BaseStore<DevToolStoreData>;
