+++ src/background/stores/dev-tools-store.ts
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 import { StoreNames } from '../../common/stores/store-names';
-import { DevToolStoreData } from '../../common/types/store-data/idev-tool-state';
+import { DevToolStoreData } from '../../common/types/store-data/dev-tool-store-data';
 import { DevToolActions } from '../actions/dev-tools-actions';
 import { BaseStoreImpl } from './base-store-impl';
