+++ src/injected/frame-url-search-initiator.ts
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 import { BaseStore } from '../common/base-store';
-import { DevToolStoreData } from '../common/types/store-data/idev-tool-state';
+import { DevToolStoreData } from '../common/types/store-data/dev-tool-store-data';
 import { FrameUrlFinder } from './frame-url-finder';
 
 export class FrameUrlSearchInitiator {
