+++ src/injected/layered-details-dialog-component.tsx
@@ -8,7 +8,7 @@ import { BaseStore } from '../common/base-store';
 import { FeatureFlags } from '../common/feature-flags';
 import { DevToolActionMessageCreator } from '../common/message-creators/dev-tool-action-message-creator';
 import { NamedFC } from '../common/react/named-fc';
-import { DevToolStoreData } from '../common/types/store-data/idev-tool-state';
+import { DevToolStoreData } from '../common/types/store-data/dev-tool-store-data';
 import { UserConfigurationStoreData } from '../common/types/store-data/user-configuration-store';
 import { DictionaryStringTo } from '../types/common-types';
 import { DetailsDialog, DetailsDialogDeps } from './components/details-dialog';
